requirements.txt
openai
python-dotenv
twilio
